Module rdkafka::client

Common client functionality.

In librdkafka parlance, a client is either a consumer or a producer. This module's Client type provides the functionality that is common to both consumers and producers.

Typically you will not want to construct a client directly. Construct one of the consumers in the consumer module or one of the producers in the producer modules instead.

Structs

Client
A low-level rdkafka client.

DefaultClientContext
An empty ClientContext that can be used when no customizations are needed.

NativeClient
A native rdkafka-sys client. This struct shouldn't be used directly. Use higher level Client or producers and consumers.

Traits

ClientContext
Client-level context.