Struct rdkafka::client::Client

pub struct Client<C: Context> { /* fields omitted */ }

A low level rdkafka client. This client shouldn't be used directly. The producer and consumer modules provide different producer and consumer implementations based on top of Client that can be used instead.

Methods

impl<C: Context> Client<C>

fn new(
config: &ClientConfig,
native_config: NativeClientConfig,
rd_kafka_type: RDKafkaType,
context: C
) -> KafkaResult<Client<C>>

Creates a new Client given a configuration, a client type and a context.

fn native_client(&self) -> &NativeClient

Returns a reference to the native rdkafka-sys client.

fn native_ptr(&self) -> *mut RDKafka

Returns a pointer to the native rdkafka-sys client.

fn context(&self) -> &C

Returns a reference to the context.

fn fetch_metadata(
&self,
topic: Option<&str>,
timeout_ms: i32
) -> KafkaResult<Metadata>

Returns the metadata information for the specified topic, or for all topics in the cluster if no topic is specified.

fn fetch_watermarks(
&self,
topic: &str,
partition: i32,
timeout_ms: i32
) -> KafkaResult<(i64, i64)>

Returns high and low watermark for the specified topic and partition.

fn fetch_group_list(
&self,
group: Option<&str>,
timeout_ms: i32
) -> KafkaResult<GroupList>

Returns the group membership information for the given group. If no group is specified, all groups will be returned.