Struct rdkafka::config::ClientConfig

pub struct ClientConfig {
    pub log_level: RDKafkaLogLevel,
    // some fields omitted
}

Client configuration.

Fields
log_level: RDKafkaLogLevel

The librdkafka logging level. Refer to RDKafkaLogLevel for the list of available levels.

Implementations

impl ClientConfig

pub fn new() -> ClientConfig

Creates a new empty configuration.

pub fn get(&self, key: &str) -> Option<&str>

Gets the value of a parameter in the configuration.

Returns the current value set for key, or None if no value for key exists.

Note that this method will only ever return values that were installed by a call to ClientConfig::set. To retrieve librdkafka’s default value for a parameter, build a NativeClientConfig and then call NativeClientConfig::get on the resulting object.

pub fn set<K, V>(&mut self, key: K, value: V) -> &mut ClientConfig where
K: Into<String>,
V: Into<String>,

Sets a parameter in the configuration.

If there is an existing value for key in the configuration, it is overridden with the new value.

pub fn remove<'a>(&'a mut self, key: &str) -> &'a mut ClientConfig

Removes a parameter from the configuration.

pub fn set_log_level(&mut self, log_level: RDKafkaLogLevel) -> &mut ClientConfig

Sets the log level of the client. If not specified, the log level will be calculated based on the global log level of the log crate.

pub fn create_native_config(&self) -> KafkaResult<NativeClientConfig>

Builds a native librdkafka configuration.

pub fn create<T: FromClientConfig>(&self) -> KafkaResult<T>

Uses the current configuration to create a new Consumer or Producer.

pub fn create_with_context<C, T>(&self, context: C) -> KafkaResult<T> where
C: ClientContext,
T: FromClientConfigAndContext<C>,

Uses the current configuration and the provided context to create a new Consumer or Producer.