Struct rdkafka::config::NativeClientConfig

pub struct NativeClientConfig { /* fields omitted */ }

A native rdkafka-sys client config.

Methods

impl NativeClientConfig

pub unsafe fn from_ptr(ptr: *mut RDKafkaConf) -> NativeClientConfig

Wraps a pointer to an RDKafkaConfig object and returns a new NativeClientConfig.

pub fn ptr(&self) -> *mut RDKafkaConf

Returns the pointer to the librdkafka RDKafkaConf structure.

pub fn ptr_move(self) -> *mut RDKafkaConf

Returns the pointer to the librdkafka RDKafkaConf structure. This method should be used when the native pointer is intended to be moved. The destructor won't be executed automatically; the caller should take care of deallocating the resource when no longer needed.