Struct rdkafka::config::NativeTopicConfig

pub struct NativeTopicConfig { /* fields omitted */ }

A native rdkafka-sys client config.

Methods

impl NativeTopicConfig

fn from_ptr(ptr: *mut RDKafkaTopicConf) -> NativeTopicConfig

Wraps a pointer to an RDKafkaTopicConf object and returns a new NativeTopicConfig.

fn ptr(&self) -> *mut RDKafkaTopicConf

Returns the pointer to the librdkafka RDKafkaTopicConf structure.

fn ptr_move(self) -> *mut RDKafkaTopicConf

Returns the pointer to the librdkafka RDKafkaTopicConf structure. This method should be used when the native pointer is intended to be moved. The destructor won't be executed automatically; the caller should take care of deallocating the resource when no longer needed.