Struct rdkafka::consumer::stream_consumer::StreamConsumer

#[must_use = "Consumer polling thread will stop immediately if unused"]pub struct StreamConsumer<C: ConsumerContext + 'static = DefaultConsumerContext> { /* fields omitted */ }

A Kafka consumer providing a futures::Stream interface.

This consumer doesn't need to be polled explicitly since awaiting the stream returned by StreamConsumer::start will implicitly poll the consumer.

Implementations

impl<C: ConsumerContext> StreamConsumer<C>

pub fn start(&self) -> MessageStream<C, TokioRuntime>

Starts the stream consumer with default configuration (100ms polling interval and no NoMessageReceived notifications).

Note: this method must be called from within the context of a Tokio runtime.

pub fn start_with(
&self,
poll_interval: Duration,
no_message_error: bool
) -> MessageStream<C, TokioRuntime>

Starts the stream consumer with the specified poll interval.

If no_message_error is set to true, the returned MessageStream will yield an error of type KafkaError::NoMessageReceived every time the poll interval is reached and no message has been received.

pub fn start_with_runtime<R>(
&self,
poll_interval: Duration,
no_message_error: bool
) -> MessageStream<C, R> where
R: AsyncRuntime,

Like StreamConsumer::start_with, but with a customizable asynchronous runtime.

See the AsyncRuntime trait for the details on the interface the runtime must satisfy.