Struct rdkafka::groups::GroupList

pub struct GroupList(_);

List of groups. This structure wraps the pointer returned by rdkafka-sys, and deallocates all the native resources when dropped.

Methods

impl GroupList

pub unsafe fn from_ptr(ptr: *const RDKafkaGroupList) -> GroupList

Creates a new group list given a pointer to the native rdkafka-sys group list structure.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn groups(&self) -> &[GroupInfo]

Returns all the groups in the list.