Module rdkafka::message

Store and manipulate Kafka messages.

Structs

BorrowedHeaders

A zero-copy collection of Kafka message headers.

BorrowedMessage

A zero-copy Kafka message.

OwnedHeaders

A collection of Kafka message headers that owns its backing data.

OwnedMessage

A Kafka message that owns its backing data.

Enums

Timestamp

Timestamp of a Kafka message.

Traits

FromBytes

A cheap conversion from a byte slice to typed data.

Headers

A generic representation of Kafka message headers.

Message

A generic representation of a Kafka message.

ToBytes

A cheap conversion from typed data to a byte slice.

Type Definitions

DeliveryResult

The result of a message production.