Struct rdkafka::message::BorrowedMessage

pub struct BorrowedMessage<'a> { /* fields omitted */ }

A zero-copy Kafka message.

The content of the message is stored in the receiving buffer of the consumer or the producer. As such, BorrowedMessage cannot outlive the consumer or producer it belongs to.

Consumers

BorrowedMessages coming from consumers are removed from the consumer buffer once they are dropped. Holding references to too many messages will cause the memory of the consumer to fill up and the consumer to block until some of the BorrowedMessages are dropped.

Conversion to owned

To transform a BorrowedMessage into a OwnedMessage, use the detach method.

Methods

impl<'a> BorrowedMessage<'a>

pub fn ptr(&self) -> *mut RDKafkaMessage

Returns a pointer to the RDKafkaMessage.

pub fn topic_ptr(&self) -> *mut RDKafkaTopic

Returns a pointer to the message's RDKafkaTopic

pub fn key_len(&self) -> usize

Returns the length of the key field of the message.

pub fn payload_len(&self) -> usize

Returns the length of the payload field of the message.

pub fn detach(&self) -> OwnedMessage

Clones the content of the BorrowedMessage and returns an OwnedMessage, that can outlive the consumer. This operation requires memory allocation and can be expensive.