Struct rdkafka::message::OwnedHeaders

pub struct OwnedHeaders { /* fields omitted */ }

Owned message headers

Kafka supports associating an array of key-value pairs to every message, called message headers. The OwnedHeaders can be used to create the desired headers and to pass them to the producer. See also [BorrowedHeaders].

Methods

impl OwnedHeaders

pub fn new() -> OwnedHeaders

Create a new OwnedHeaders struct with initial capacity 5.

pub fn new_with_capacity(initial_capacity: usize) -> OwnedHeaders

Create a new OwnedHeaders struct with the desired initial capacity. The structure is automatically resized as more headers are added.

pub fn add<V: ToBytes + ?Sized>(self, name: &str, value: &V) -> OwnedHeaders

Add a new header to the structure.

pub fn as_borrowed(&self) -> &BorrowedHeaders

Generate a read-only [BorrowedHeaders] reference.