Struct rdkafka::metadata::Metadata

pub struct Metadata(_);

Metadata container. This structure wraps the metadata pointer returned by rdkafka-sys, and deallocates all the native resources when dropped.

Methods

impl Metadata

pub unsafe fn from_ptr(ptr: *const RDKafkaMetadata) -> Metadata

Creates a new Metadata container given a pointer to the native rdkafka-sys metadata.

pub fn orig_broker_id(&self) -> i32

Returns the id of the broker originating this metadata.

pub fn orig_broker_name(&self) -> &str

Returns the hostname of the broker originating this metadata.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn brokers(&self) -> &[MetadataBroker]

Returns the metadata information for all the brokers in the cluster.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn topics(&self) -> &[MetadataTopic]

Returns the metadata information for all the topics in the cluster.