Struct rdkafka::metadata::MetadataPartition

pub struct MetadataPartition(_);

Partition metadata information.

Methods

impl MetadataPartition

pub fn id(&self) -> i32

Returns the id of the partition.

pub fn leader(&self) -> i32

Returns the broker id of the leader broker for the partition.

pub fn error(&self) -> Option<RDKafkaRespErr>

Returns the metadata error for the partition, or None if there is no error.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn replicas(&self) -> &[i32]

Returns the broker ids of the replicas.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn isr(&self) -> &[i32]

Returns the broker ids of the in sync replicas.