Struct rdkafka::metadata::MetadataTopic

pub struct MetadataTopic(_);

Topic metadata information.

Methods

impl MetadataTopic

pub fn name(&self) -> &str

Returns the name of the topic.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn partitions(&self) -> &[MetadataPartition]

Returns the partition metadata information for all the partitions.

pub fn error(&self) -> Option<RDKafkaRespErr>

Returns the metadata error, or None if there was no error.