Trait rdkafka::producer::base_producer::ProducerContext

pub trait ProducerContext: ClientContext {
    type DeliveryOpaque: IntoOpaque;
    fn delivery(
        &self, 
        delivery_result: &DeliveryResult, 
        delivery_opaque: Self::DeliveryOpaque
    );
}

A ProducerContext is a Context specific for producers. It can be used to store user-specified callbacks, such as delivery.

Associated Types

type DeliveryOpaque: IntoOpaque

A DeliveryOpaque is a user-defined structure that will be passed to the producer when producing a message, and returned to the delivery method once the message has been delivered, or failed to.

Required Methods

fn delivery(
&self,
delivery_result: &DeliveryResult,
delivery_opaque: Self::DeliveryOpaque
)

This method will be called once the message has been delivered (or failed to). The DeliveryOpaque will be the one provided by the user when calling send.