Module rdkafka::producer::future_producer

High-level, futures-enabled Kafka producer.

See the FutureProducer for details.

Structs

DeliveryFuture
A Future wrapping the result of the message production.

FutureProducer
A producer that returns a Future for every message being produced.

FutureProducerContext
The ProducerContext used by the FutureProducer.

FutureRecord
A record for the future producer.

Type Definitions

OwnedDeliveryResult
Represents the result of message production as performed from the FutureProducer.