Struct rdkafka::producer::future_producer::DeliveryFuture

pub struct DeliveryFuture { /* fields omitted */ }

A Future wrapping the result of the message production.

Once completed, the future will contain an OwnedDeliveryResult with information on the delivery status of the message. If the producer is dropped before the delivery status is received, the future will instead resolve with oneshot::Canceled.