Struct rdkafka::producer::BaseProducerTopic

pub struct BaseProducerTopic<C: ProducerContext> { /* fields omitted */ }

Represents a Kafka topic with an associated BaseProducer.

Methods

impl<C: ProducerContext> BaseProducerTopic<C>

fn new(producer: BaseProducer<C>,
name: &str,
topic_config: &TopicConfig)
-> KafkaResult<BaseProducerTopic<C>>

Creates the BaseProducerTopic.

fn name(&self) -> String

Get topic's name

fn send_copy<P, K>(&self,
partition: Option<i32>,
payload: Option<&P>,
key: Option<&K>,
delivery_context: Option<Box<C::DeliveryContext>>,
timestamp: Option<i64>)
-> KafkaResult<()> where K: ToBytes, P: ToBytes

Sends a copy of the payload and key provided to the specified topic. When no partition is specified the underlying Kafka library picks a partition based on the key. If no key is specified, a random partition will be used.