Struct rdkafka::producer::FutureProducer

pub struct FutureProducer<C = DefaultClientContext, R = DefaultRuntime> where
    C: ClientContext + 'static,  { /* fields omitted */ }

A producer that returns a Future for every message being produced.

Since message production in rdkafka is asynchronous, the caller cannot immediately know if the delivery of the message was successful or not. The FutureProducer provides this information in a Future, which will be completed once the information becomes available.

This producer has an internal polling thread and as such it doesn’t need to be polled. It can be cheaply cloned to get a reference to the same underlying producer. The internal polling thread will be terminated when the FutureProducer goes out of scope.

Implementations

impl<C, R> FutureProducer<C, R> where
C: ClientContext + 'static,
R: AsyncRuntime,

pub async fn send<K: ?Sized, P: ?Sized, T>(
&self,
record: FutureRecord<'_, K, P>,
queue_timeout: T
) -> OwnedDeliveryResult where
K: ToBytes + ?Sized,
P: ToBytes + ?Sized,
T: Into<Timeout>,

Sends a message to Kafka, returning the result of the send.

The queue_timeout parameter controls how long to retry for if the librdkafka producer queue is full. Set it to Timeout::Never to retry forever or Timeout::After(0) to never block. If the timeout is reached and the queue is still full, an RDKafkaErrorCode::QueueFull error will be reported in the OwnedDeliveryResult.

Keep in mind that queue_timeout only applies to the first phase of the send operation. Once the message is queued, the underlying librdkafka client has separate timeout parameters that apply, like delivery.timeout.ms.

See also the FutureProducer::send_result method, which will not retry the queue operation if the queue is full.

pub fn send_result<'a, K: ?Sized, P: ?Sized>(
&self,
record: FutureRecord<'a, K, P>
) -> Result<DeliveryFuture, (KafkaError, FutureRecord<'a, K, P>)> where
K: ToBytes + ?Sized,
P: ToBytes + ?Sized,

Like FutureProducer::send, but if enqueuing fails, an error will be returned immediately, alongside the FutureRecord provided.

pub fn poll<T: Into<Timeout>>(&self, timeout: T)

Polls the internal producer.

This is not normally required since the FutureProducer has a thread dedicated to calling poll regularly.