Struct rdkafka::producer::FutureProducerContext

pub struct FutureProducerContext<C: Context + 'static> { /* fields omitted */ }

The ProducerContext used by the FutureProducer. This context will use a Future as its DeliveryContext and will complete the future when the message is delivered (or failed to).