Struct rdkafka::statistics::Broker

pub struct Broker {

    pub name: String,
    pub nodeid: i32,
    pub nodename: String,
    pub source: String,
    pub state: String,
    pub stateage: i64,
    pub outbuf_cnt: i64,
    pub outbuf_msg_cnt: i64,
    pub waitresp_cnt: i64,
    pub waitresp_msg_cnt: i64,
    pub tx: u64,
    pub txbytes: u64,
    pub txerrs: u64,
    pub txretries: u64,
    pub txidle: i64,
    pub req_timeouts: u64,
    pub rx: u64,
    pub rxbytes: u64,
    pub rxerrs: u64,
    pub rxcorriderrs: u64,
    pub rxpartial: u64,
    pub rxidle: i64,
    pub req: HashMap<String, i64>,
    pub zbuf_grow: u64,
    pub buf_grow: u64,
    pub wakeups: Option<u64>,
    pub connects: Option<i64>,
    pub disconnects: Option<i64>,
    pub int_latency: Option<Window>,
    pub outbuf_latency: Option<Window>,
    pub rtt: Option<Window>,
    pub throttle: Option<Window>,
    pub toppars: HashMap<String, TopicPartition>,

}

Per-broker statistics.

Fields
name: String

The broker hostname, port, and ID, in the form HOSTNAME:PORT/ID.

nodeid: i32

The broker ID (-1 for bootstraps).

nodename: String

The broker hostname and port.

source: String

The broker source (learned, configured, internal, or logical).

state: String

The broker state (INIT, DOWN, CONNECT, AUTH, APIVERSION_QUERY, AUTH_HANDSHAKE, UP, UPDATE).

stateage: i64

The time since the last broker state change, in microseconds.

outbuf_cnt: i64

The number of requests awaiting transmission to the broker.

outbuf_msg_cnt: i64

The number of messages awaiting transmission to the broker.

waitresp_cnt: i64

The number of requests in-flight to the broker that are awaiting a response.

waitresp_msg_cnt: i64

The number of messages in-flight to the broker that are awaiting a response.

tx: u64

The total number of requests sent to the broker.

txbytes: u64

The total number of bytes sent to the broker.

txerrs: u64

The total number of transmission errors.

txretries: u64

The total number of request retries.

txidle: i64

Microseconds since last socket send, or -1 if no sends yet for the current connection.

req_timeouts: u64

The total number of requests that timed out.

rx: u64

The total number of responses received from the broker.

rxbytes: u64

The total number of bytes received from the broker.

rxerrs: u64

The total number of receive errors.

rxcorriderrs: u64

The number of unmatched correlation IDs in response, typically for timed out requests.

rxpartial: u64

The total number of partial message sets received. The broker may return partial responses if the full message set could not fit in the remaining fetch response size.

rxidle: i64

Microseconds since last socket receive, or -1 if no receives yet for the current connection.

req: HashMap<String, i64>

Request type counters. The object key is the name of the request type and the value is the number of requests of that type that have been sent.

zbuf_grow: u64

The total number of decompression buffer size increases.

buf_grow: u64

The total number of buffer size increases (deprecated and unused).

wakeups: Option<u64>

The number of broker thread poll wakeups.

connects: Option<i64>

The number of connection attempts, including successful and failed attempts, and name resolution failures.

disconnects: Option<i64>

The number of disconnections, whether triggered by the broker, the network, the load balancer, or something else.

int_latency: Option<Window>

Rolling window statistics for the internal producer queue latency, in microseconds.

outbuf_latency: Option<Window>

Rolling window statistics for the internal request queue latency, in microseconds.

This is the time between when a request is enqueued on the transmit (outbuf) queue and the time the request is written to the TCP socket. Additional buffering and latency may be incurred by the TCP stack and network.

rtt: Option<Window>

Rolling window statistics for the broker latency/round-trip time, in microseconds.

throttle: Option<Window>

Rolling window statistics for the broker throttling time, in milliseconds.

toppars: HashMap<String, TopicPartition>

The partitions that are handled by this broker handle.