Struct rdkafka::statistics::Partition

pub struct Partition {
    pub partition: i32,
    pub broker: i32,
    pub leader: i32,
    pub desired: bool,
    pub unknown: bool,
    pub msgq_cnt: i64,
    pub msgq_bytes: i64,
    pub xmit_msgq_cnt: i64,
    pub xmit_msgq_bytes: i64,
    pub fetchq_cnt: i64,
    pub fetchq_size: i64,
    pub fetch_state: String,
    pub query_offset: i64,
    pub next_offset: i64,
    pub app_offset: i64,
    pub stored_offset: i64,
    pub committed_offset: i64,
    pub eof_offset: i64,
    pub lo_offset: i64,
    pub hi_offset: i64,
    pub ls_offset: i64,
    pub consumer_lag: i64,
    pub txmsgs: i64,
    pub txbytes: i64,
    pub rxmsgs: i64,
    pub rxbytes: i64,
    pub msgs: i64,
    pub rx_ver_drops: i64,
    pub msgs_inflight: i64,
    pub next_ack_seq: i64,
    pub next_err_seq: i64,
    pub acked_msgid: i64,
}

Per-partition statistics.

Fields
partition: i32

The partition ID.

broker: i32

The ID of the broker from which messages are currently being fetched.

leader: i32

The broker ID of the leader.

desired: bool

Whether the partition is explicitly desired by the application.

unknown: bool

Whether the partition is not seen in the topic metadata from the broker.

msgq_cnt: i64

The number of messages waiting to be produced in the first-level queue.

msgq_bytes: i64

The number of bytes waiting to be produced in the first-level queue.

xmit_msgq_cnt: i64

The number of messages ready to be produced in the transmit queue.

xmit_msgq_bytes: i64

The number of bytes ready to be produced in the transmit queue.

fetchq_cnt: i64

The number of prefetched messages in the fetch queue.

fetchq_size: i64

The number of bytes in the fetch queue.

fetch_state: String

The consumer fetch state for this partition (none, stopping, stopped, offset-query, offset-wait, active).

query_offset: i64

The current/last logical offset query.

next_offset: i64

The next offset to fetch.

app_offset: i64

The offset of the last message passed to the application, plus one.

stored_offset: i64

The offset to be committed.

committed_offset: i64

The last committed offset.

eof_offset: i64

The last offset for which partition EOF was signaled.

lo_offset: i64

The low watermark offset on the broker.

hi_offset: i64

The high watermark offset on the broker.

ls_offset: i64

The last stable offset on the broker.

consumer_lag: i64

The difference between hi_offset and max(app_offset, committed_offset).

txmsgs: i64

The total number of messages transmitted (produced).

txbytes: i64

The total number of bytes transmitted (produced).

rxmsgs: i64

The total number of messages consumed, not included ignored messages.

rxbytes: i64

The total bytes consumed.

msgs: i64

The total number of messages received, for consumers, or the total number of messages produced, for producers.

rx_ver_drops: i64

The number of dropped outdated messages.

msgs_inflight: i64

The current number of messages in flight to or from the broker.

next_ack_seq: i64

The next expected acked sequence number, for idempotent producers.

next_err_seq: i64

The next expected errored sequence number, for idempotent producers.

acked_msgid: i64

The last acked internal message ID, for idempotent producers.