Enum rdkafka::topic_partition_list::Offset

pub enum Offset {
    Beginning,
    End,
    Stored,
    Invalid,
    Offset(i64),
    OffsetTail(i64),
}

A Kafka offset.

Variants

Beginning

Start consuming from the beginning of the partition.

End

Start consuming from the end of the partition.

Stored

Start consuming from the stored offset.

Invalid

Offset not assigned or invalid.

Offset(i64)

A specific offset to consume from.

Note that while the offset is a signed integer, negative offsets will be rejected when passed to librdkafka.

OffsetTail(i64)

An offset relative to the end of the partition.

Note that while the offset is a signed integer, negative offsets will be rejected when passed to librdkafka.

Implementations

impl Offset

pub fn from_raw(raw_offset: i64) -> Offset

Converts the integer representation of an offset used by librdkafka to an Offset.

pub fn to_raw(self) -> Option<i64>

Converts the Offset to the internal integer representation used by librdkafka.

Returns None if the offset cannot be represented in librdkafka's internal representation.