Struct rdkafka::topic_partition_list::TopicPartitionList

pub struct TopicPartitionList { /* fields omitted */ }

A structure to store and manipulate a list of topics and partitions with optional offsets.

Implementations

impl TopicPartitionList

pub fn new() -> TopicPartitionList

Creates a new empty list with default capacity.

pub fn with_capacity(capacity: usize) -> TopicPartitionList

Creates a new empty list with the specified capacity.

pub fn from_topic_map(
topic_map: &HashMap<(String, i32), Offset>
) -> KafkaResult<TopicPartitionList>

Given a topic map, generates a new TopicPartitionList.

pub fn ptr(&self) -> *mut RDKafkaTopicPartitionList

Returns the pointer to the internal librdkafka structure.

pub fn count(&self) -> usize

Returns the number of elements in the list.

pub fn capacity(&self) -> usize

Returns the capacity of the list.

pub fn add_topic_unassigned<'a>(
&'a mut self,
topic: &str
) -> TopicPartitionListElem<'a>

Adds a topic with unassigned partitions to the list.

pub fn add_partition<'a>(
&'a mut self,
topic: &str,
partition: i32
) -> TopicPartitionListElem<'a>

Adds a topic and partition to the list.

pub fn add_partition_range(
&mut self,
topic: &str,
start_partition: i32,
stop_partition: i32
)

Adds a topic and partition range to the list.

pub fn set_partition_offset(
&mut self,
topic: &str,
partition: i32,
offset: Offset
) -> KafkaResult<()>

Sets the offset for an already created topic partition. It will fail if the topic partition isn’t in the list.

pub fn add_partition_offset(
&mut self,
topic: &str,
partition: i32,
offset: Offset
) -> KafkaResult<()>

Adds a topic and partition to the list, with the specified offset.

pub fn find_partition(
&self,
topic: &str,
partition: i32
) -> Option<TopicPartitionListElem<'_>>

Given a topic name and a partition number, returns the corresponding list element.

pub fn set_all_offsets(&mut self, offset: Offset) -> Result<(), KafkaError>

Sets all partitions in the list to the specified offset.

pub fn elements(&self) -> Vec<TopicPartitionListElem<'_>>

Returns all the elements of the list.

pub fn elements_for_topic<'a>(
&'a self,
topic: &str
) -> Vec<TopicPartitionListElem<'a>>

Returns all the elements of the list that belong to the specified topic.

pub fn to_topic_map(&self) -> HashMap<(String, i32), Offset>

Returns a hashmap-based representation of the list.