Trait rdkafka::util::IntoOpaque

pub trait IntoOpaque: Send + Sync {
    fn as_ptr(&self) -> *mut c_void;

    unsafe fn from_ptr(_: *mut c_void) -> Self;
}

A trait for the conversion of Rust data to raw pointers. This conversion is used to pass opaque objects to the C library and vice versa.

Required Methods

fn as_ptr(&self) -> *mut c_void

Converts the object into a raw pointer.

unsafe fn from_ptr(_: *mut c_void) -> Self

Converts the raw pointer back to the original Rust object.