Module rdkafka::client

Common client functionalities.

Structs

Client
A low level rdkafka client. This client shouldn't be used directly. The producer and consumer modules provide different producer and consumer implementations based on top of Client that can be used instead.

EmptyContext
An empty context that can be used when no context is needed.

NativeClient
A native rdkafka-sys client. This struct shouldn't be used directly. Use higher level Client or producers and consumers.

Traits

Context
A Context is an object that can store user-defined data and on which callbacks can be defined. Refer to the list of methods to see which callbacks can currently be overridden. The context must be thread safe, and might be owned by multiple threads.

Functions

native_error_cb⚠

native_log_cb⚠

native_stats_cb⚠