Struct rdkafka::client::Client

pub struct Client<C: ClientContext = DefaultClientContext> { /* fields omitted */ }

A low level rdkafka client. This client shouldn't be used directly. The producer and consumer modules provide different producer and consumer implementations based on top of Client that can be used instead.

Methods

impl<C: ClientContext> Client<C>

pub fn new(
config: &ClientConfig,
native_config: NativeClientConfig,
rd_kafka_type: RDKafkaType,
context: C
) -> KafkaResult<Client<C>>

Creates a new Client given a configuration, a client type and a context.

pub fn native_client(&self) -> &NativeClient

Returns a reference to the native rdkafka-sys client.

pub fn native_ptr(&self) -> *mut RDKafka

Returns a pointer to the native rdkafka-sys client.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn context(&self) -> &C

Returns a reference to the context.

pub fn fetch_metadata<T: Into<Option<Duration>>>(
&self,
topic: Option<&str>,
timeout: T
) -> KafkaResult<Metadata>

Returns the metadata information for the specified topic, or for all topics in the cluster if no topic is specified.

pub fn fetch_watermarks<T: Into<Option<Duration>>>(
&self,
topic: &str,
partition: i32,
timeout: T
) -> KafkaResult<(i64, i64)>

Returns high and low watermark for the specified topic and partition.

pub fn fetch_group_list<T: Into<Option<Duration>>>(
&self,
group: Option<&str>,
timeout: T
) -> KafkaResult<GroupList>

Returns the group membership information for the given group. If no group is specified, all groups will be returned.