Module rdkafka::config

Producer and consumer configuration.

C library configuration

The Rust library will forward all the configuration to the C library. The most frequently used parameters are listed here.

Frequently used parameters

For producer-specific and consumer-specific parameters check the producer and consumer modules documentation. The full list of available parameters is available in the librdkafka documentation.

client.id: Client identifier. Default: rdkafka.

bootstrap.servers: Initial list of brokers as a CSV list of broker host or host:port. Default: empty.

message.max.bytes: Maximum message size. Default: 1000000.

debug: A comma-separated list of debug contexts to enable. Use ‘all’ to print all the debugging information. Default: empty (off).

statistics.interval.ms: how often the statistic callback specified in the ClientContext will be called. Default: 0 (disabled).

Structs

ClientConfig

Client configuration.

NativeClientConfig

A native rdkafka-sys client config.

Enums

RDKafkaLogLevel

The log levels supported by librdkafka.

Traits

FromClientConfig

Create a new client based on the provided configuration.

FromClientConfigAndContext

Create a new client based on the provided configuration and context.