Module rdkafka::config

Producer and consumer configuration.

C library configuration

The Rust library will forward all the configuration to the C library. The most frequently used parameters are listed here.

Frequently used parameters

For producer-specific and consumer-specific parameters check the producer and consumer modules documentation. The full list of available parameters is available in the librdkafka documentation.

client.id (rdkafka): Client identifier.

bootstrap.servers: Initial list of brokers as a CSV list of broker host or host:port.

message.max.bytes (1000000): Maximum message size.

debug: A comma-separated list of debug contexts to enable. Use 'all' to print all the debugging information.

statistics.interval.ms (0 - disabled): how often the statistic callback specified in the Context will be called.

Structs

ClientConfig
Client configuration.

NativeClientConfig
A native rdkafka-sys client config.

Enums

RDKafkaLogLevel
The log levels supported by librdkafka.

Traits

FromClientConfig
Create a new client based on the provided configuration.

FromClientConfigAndContext
Create a new client based on the provided configuration and context.