Trait rdkafka::config::FromClientConfigAndContext

pub trait FromClientConfigAndContext<C: ClientContext>: Sized {
    pub fn from_config_and_context(_: &ClientConfig, _: C) -> KafkaResult<Self>;
}

Create a new client based on the provided configuration and context.

Required methods

pub fn from_config_and_context(_: &ClientConfig, _: C) -> KafkaResult<Self>

Creates a client from a client configuration and a client context.

Loading content...