Module rdkafka::consumer

Kafka consumers.

Modules

base_consumer
Low-level consumers.

stream_consumer
High-level consumers with a Stream interface.

Structs

BaseConsumer
A low-level consumer that requires manual polling.

ConsumerGroupMetadata
Consumer group metadata.

DefaultConsumerContext
An inert ConsumerContext that can be used when no customizations are needed.

MessageStream
A stream of messages from a StreamConsumer.

StreamConsumer
A high-level consumer with a Stream interface.

Enums

CommitMode
Specifies whether a commit should be performed synchronously or asynchronously.

Rebalance
Rebalance information.

Traits

Consumer
Common trait for all consumers.

ConsumerContext
Consumer-specific context.