Module rdkafka::consumer

Base trait and common functionality for all consumers.

Re-exports

pub use self::base_consumer::BaseConsumer;

pub use self::stream_consumer::MessageStream;

pub use self::stream_consumer::StreamConsumer;

Modules

base_consumer
Low level consumer wrapper.

stream_consumer
Stream-based consumer implementation.

Structs

DefaultConsumerContext
An empty consumer context that can be user when no context is needed.

Enums

CommitMode
Specifies if the commit should be performed synchronously or asynchronously.

Rebalance
Rebalance information.

Traits

Consumer
Common trait for all consumers.

ConsumerContext
Consumer specific Context. This user-defined object can be used to provide custom callbacks to consumer events. Refer to the list of methods to check which callbacks can be specified.