Struct rdkafka::consumer::base_consumer::BaseConsumer

pub struct BaseConsumer<C: ConsumerContext = DefaultConsumerContext> { /* fields omitted */ }

Low level wrapper around the librdkafka consumer. This consumer requires to be periodically polled to make progress on rebalance, callbacks and to receive messages.

Methods

impl<C: ConsumerContext> BaseConsumer<C>

pub fn poll<T: Into<Option<Duration>>>(
&self,
timeout: T
) -> Option<KafkaResult<BorrowedMessage>>

Polls the consumer for new messages.

It won't block for more than the specified timeout. Use zero Duration for non-blocking call. With no timeout it blocks until an event is received.

This method should be called at regular intervals, even if no message is expected, to serve any queued callbacks waiting to be called. This is especially important for automatic consumer rebalance, as the rebalance function will be executed by the thread calling the poll() function.

Lifetime

The returned message lives in the memory of the consumer and cannot outlive it.

Important traits for Iter<'a, C>
impl<'a, C: ConsumerContext + 'a> Iterator for Iter<'a, C> type Item = KafkaResult<BorrowedMessage<'a>>;

pub fn iter(&self) -> Iter<C>

Returns an iterator over the available messages.

It repeatedly calls poll with no timeout.

Note that it's also possible to iterate over the consumer directly.

Examples

All these are equivalent and will receive messages without timing out.

loop {
  let message = consumer.poll(None);
  // Handle the message
}

for message in consumer.iter() {
  // Handle the message
}

for message in &consumer {
  // Handle the message
}