Struct rdkafka::consumer::base_consumer::BaseConsumer

pub struct BaseConsumer<C: ConsumerContext> { /* fields omitted */ }

Low level wrapper around the librdkafka consumer. This consumer requires to be periodically polled to make progress on rebalance, callbacks and to receive messages.

Methods

impl<C: ConsumerContext> BaseConsumer<C>

fn poll_raw(&self, timeout_ms: i32) -> KafkaResult<Option<*mut RDKafkaMessage>>

Polls the consumer for messages and returns a pointer to the native rdkafka-sys struct. This method is for internal use only.

fn poll(&self, timeout_ms: i32) -> KafkaResult<Option<BorrowedMessage>>

Polls the consumer for new messages. It won't block more than the specified timeout. Set the timeout to 0 to make the call non-blocking, or to -1 to block until an event is received. This method should be called at regular intervals, even if no message is expected, to serve any queued callbacks waiting to be called. This is especially important for automatic consumer rebalance, as the rebalance function will be executed by the thread calling the poll() function.