Struct rdkafka::consumer::stream_consumer::StreamConsumer

#[must_use = "Consumer polling thread will stop immediately if unused"]

pub struct StreamConsumer<C: ConsumerContext + 'static> { /* fields omitted */ }

A Consumer with an associated polling thread. This consumer doesn't need to be polled and it will return all consumed messages as a Stream. Due to the asynchronous nature of the stream, some messages might be consumed by the consumer without being processed on the other end of the stream. If auto commit is used, it might cause message loss after consumer restart. Manual offset storing should be used, see the store_offset function on Consumer.

Methods

impl<C: ConsumerContext> StreamConsumer<C>

fn start(&self) -> MessageStream<C>

Starts the StreamConsumer with default configuration (100ms polling interval and no NoMessageReceived notifications).

fn start_with(
&self,
poll_interval: Duration,
no_message_error: bool
) -> MessageStream<C>

Starts the StreamConsumer with the specified poll interval. Additionally, if no_message_error is set to true, it will return an error of type KafkaError::NoMessageReceived every time the poll interval is reached and no message has been received.

fn stop(&self)

Stops the StreamConsumer, blocking the caller until the internal consumer has been stopped.