Struct rdkafka::consumer::stream_consumer::StreamConsumer

#[must_use = "Consumer polling thread will stop immediately if unused"]

pub struct StreamConsumer<C: ConsumerContext + 'static> { /* fields omitted */ }

A Consumer with an associated polling thread. This consumer doesn't need to be polled and it will return all consumed messages as a Stream. Due to the asynchronous nature of the stream, some messages might be consumed by the consumer without being processed on the other end of the stream. If auto commit is used, it might cause message loss after consumer restart. Manual offset commit should be used instead.

Methods

impl<C: ConsumerContext> StreamConsumer<C>

fn start(&mut self) -> MessageStream

Starts the StreamConsumer, returning a MessageStream.

fn stop(&mut self)

Stops the StreamConsumer, blocking the caller until the internal consumer has been stopped.