Struct rdkafka::consumer::stream_consumer::StreamConsumerContext

pub struct StreamConsumerContext<C> where
    C: ConsumerContext,  { /* fields omitted */ }

A consumer context wrapper for a stream consumer.

This context will automatically wake up the message stream when new data is available.

This type is not intended to be used directly. The construction of a StreamConsumer automatically wraps the underlying context in this type.