Struct rdkafka::consumer::stream_consumer::StreamConsumerContext

pub struct StreamConsumerContext<C: ConsumerContext + 'static> { /* fields omitted */ }

The ConsumerContext used by the StreamConsumer. This context will automatically wake up the message stream when new data is available.

This type is not intended to be used directly. It will be automatically created by the StreamConsumer when necessary.