Trait rdkafka::consumer::Consumer

pub trait Consumer<C = DefaultConsumerContext> where
    C: ConsumerContext, {
    pub fn client(&self) -> &Client<C>;

    pub fn group_metadata(&self) -> Option<ConsumerGroupMetadata>;

    pub fn subscribe(&self, topics: &[&str]) -> KafkaResult<()>;

    pub fn unsubscribe(&self);

    pub fn assign(&self, assignment: &TopicPartitionList) -> KafkaResult<()>;

    pub fn seek<T: Into<Timeout>>(
        &self, 
        topic: &str, 
        partition: i32, 
        offset: Offset, 
        timeout: T
    ) -> KafkaResult<()>;

    pub fn commit(
        &self, 
        topic_partition_list: &TopicPartitionList, 
        mode: CommitMode
    ) -> KafkaResult<()>;

    pub fn commit_consumer_state(&self, mode: CommitMode) -> KafkaResult<()>;

    pub fn commit_message(
        &self, 
        message: &BorrowedMessage<'_>, 
        mode: CommitMode
    ) -> KafkaResult<()>;

    pub fn store_offset(&self, message: &BorrowedMessage<'_>) -> KafkaResult<()>;

    pub fn store_offsets(&self, tpl: &TopicPartitionList) -> KafkaResult<()>;

    pub fn subscription(&self) -> KafkaResult<TopicPartitionList>;

    pub fn assignment(&self) -> KafkaResult<TopicPartitionList>;

    pub fn committed<T>(&self, timeout: T) -> KafkaResult<TopicPartitionList>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn committed_offsets<T>(
        &self, 
        tpl: TopicPartitionList, 
        timeout: T
    ) -> KafkaResult<TopicPartitionList>
    where
        T: Into<Timeout>;

    pub fn offsets_for_timestamp<T>(
        &self, 
        timestamp: i64, 
        timeout: T
    ) -> KafkaResult<TopicPartitionList>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn offsets_for_times<T>(
        &self, 
        timestamps: TopicPartitionList, 
        timeout: T
    ) -> KafkaResult<TopicPartitionList>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn position(&self) -> KafkaResult<TopicPartitionList>;

    pub fn fetch_metadata<T>(
        &self, 
        topic: Option<&str>, 
        timeout: T
    ) -> KafkaResult<Metadata>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn fetch_watermarks<T>(
        &self, 
        topic: &str, 
        partition: i32, 
        timeout: T
    ) -> KafkaResult<(i64, i64)>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn fetch_group_list<T>(
        &self, 
        group: Option<&str>, 
        timeout: T
    ) -> KafkaResult<GroupList>
    where
        T: Into<Timeout>,
        Self: Sized;

    pub fn pause(&self, partitions: &TopicPartitionList) -> KafkaResult<()>;

    pub fn resume(&self, partitions: &TopicPartitionList) -> KafkaResult<()>;

    pub fn context(&self) -> &Arc<C> { ... }
}

Common trait for all consumers.

Note about object safety

Doing type erasure on consumers is expected to be rare (eg. Box<dyn Consumer>). Therefore, the API is optimised for the case where a concrete type is available. As a result, some methods are not available on trait objects, since they are generic.

Required methods

pub fn client(&self) -> &Client<C>

Returns the Client underlying this consumer.

pub fn group_metadata(&self) -> Option<ConsumerGroupMetadata>

Returns the current consumer group metadata associated with the consumer.

If the consumer was not configured with a group.id, returns None. For use with Producer::send_offsets_to_transaction.

pub fn subscribe(&self, topics: &[&str]) -> KafkaResult<()>

Subscribes the consumer to a list of topics.

pub fn unsubscribe(&self)

Unsubscribes the current subscription list.

pub fn assign(&self, assignment: &TopicPartitionList) -> KafkaResult<()>

Manually assigns topics and partitions to the consumer. If used, automatic consumer rebalance won't be activated.

pub fn seek<T: Into<Timeout>>(
&self,
topic: &str,
partition: i32,
offset: Offset,
timeout: T
) -> KafkaResult<()>

Seeks to offset for the specified topic and partition. After a successful call to seek, the next poll of the consumer will return the message with offset.

pub fn commit(
&self,
topic_partition_list: &TopicPartitionList,
mode: CommitMode
) -> KafkaResult<()>

Commits the offset of the specified message. The commit can be sync (blocking), or async. Notice that when a specific offset is committed, all the previous offsets are considered committed as well. Use this method only if you are processing messages in order.

pub fn commit_consumer_state(&self, mode: CommitMode) -> KafkaResult<()>

Commits the current consumer state. Notice that if the consumer fails after a message has been received, but before the message has been processed by the user code, this might lead to data loss. Check the "at-least-once delivery" section in the readme for more information.

pub fn commit_message(
&self,
message: &BorrowedMessage<'_>,
mode: CommitMode
) -> KafkaResult<()>

Commit the provided message. Note that this will also automatically commit every message with lower offset within the same partition.

pub fn store_offset(&self, message: &BorrowedMessage<'_>) -> KafkaResult<()>

Stores offset for this message to be used on the next (auto)commit. When using this enable.auto.offset.store should be set to false in the config.

pub fn store_offsets(&self, tpl: &TopicPartitionList) -> KafkaResult<()>

Store offsets to be used on the next (auto)commit. When using this enable.auto.offset.store should be set to false in the config.

pub fn subscription(&self) -> KafkaResult<TopicPartitionList>

Returns the current topic subscription.

pub fn assignment(&self) -> KafkaResult<TopicPartitionList>

Returns the current partition assignment.

pub fn committed<T>(&self, timeout: T) -> KafkaResult<TopicPartitionList> where
T: Into<Timeout>,
Self: Sized,

Retrieves the committed offsets for topics and partitions.

pub fn committed_offsets<T>(
&self,
tpl: TopicPartitionList,
timeout: T
) -> KafkaResult<TopicPartitionList> where
T: Into<Timeout>,

Retrieves the committed offsets for specified topics and partitions.

pub fn offsets_for_timestamp<T>(
&self,
timestamp: i64,
timeout: T
) -> KafkaResult<TopicPartitionList> where
T: Into<Timeout>,
Self: Sized,

Looks up the offsets for this consumer's partitions by timestamp.

pub fn offsets_for_times<T>(
&self,
timestamps: TopicPartitionList,
timeout: T
) -> KafkaResult<TopicPartitionList> where
T: Into<Timeout>,
Self: Sized,

Looks up the offsets for the specified partitions by timestamp.

pub fn position(&self) -> KafkaResult<TopicPartitionList>

Retrieve current positions (offsets) for topics and partitions.

pub fn fetch_metadata<T>(
&self,
topic: Option<&str>,
timeout: T
) -> KafkaResult<Metadata> where
T: Into<Timeout>,
Self: Sized,

Returns the metadata information for the specified topic, or for all topics in the cluster if no topic is specified.

pub fn fetch_watermarks<T>(
&self,
topic: &str,
partition: i32,
timeout: T
) -> KafkaResult<(i64, i64)> where
T: Into<Timeout>,
Self: Sized,

Returns the metadata information for all the topics in the cluster.

pub fn fetch_group_list<T>(
&self,
group: Option<&str>,
timeout: T
) -> KafkaResult<GroupList> where
T: Into<Timeout>,
Self: Sized,

Returns the group membership information for the given group. If no group is specified, all groups will be returned.

pub fn pause(&self, partitions: &TopicPartitionList) -> KafkaResult<()>

Pauses consumption for the provided list of partitions.

pub fn resume(&self, partitions: &TopicPartitionList) -> KafkaResult<()>

Resumes consumption for the provided list of partitions.

Loading content...
Provided methods

pub fn context(&self) -> &Arc<C>

Returns a reference to the ConsumerContext used to create this consumer.

Loading content...