Struct rdkafka::message::BorrowedMessage

pub struct BorrowedMessage<'a> { /* fields omitted */ }

A zero-copy Kafka message.

The content of the message is stored in the receiving buffer of the consumer or the producer. As such, BorrowedMessage cannot outlive the consumer or producer it belongs to.

Consumers

BorrowedMessages coming from consumers are removed from the consumer buffer once they are dropped. Holding references to too many messages will cause the memory of the consumer to fill up and the consumer to block until some of the BorrowedMessages are dropped.

Conversion to owned

To transform a BorrowedMessage into a OwnedMessage, use the detach method.

Methods

impl<'a> BorrowedMessage<'a>

unsafe fn from_consumer<C>(
ptr: *mut RDKafkaMessage,
_consumer: &'a C
) -> KafkaResult<BorrowedMessage<'a>>

Creates a new BorrowedMessage that wraps the native Kafka message pointer returned by a consumer. The lifetime of the message will be bound to the lifetime of the consumer passed as parameter. This method should only be used with messages coming from consumers. If the message contains an error, only the error is returned and the message structure is freed.

unsafe fn from_dr_callback<O>(
ptr: *mut RDKafkaMessage,
_owner: &'a O
) -> DeliveryResult<'a>

Creates a new BorrowedMessage that wraps the native Kafka message pointer returned by the delivery callback of a producer. The lifetime of the message will be bound to the lifetime of the reference passed as parameter. This method should only be used with messages coming from the delivery callback. The message will not be freed in any circumstance.

fn ptr(&self) -> *mut RDKafkaMessage

Returns a pointer to the RDKafkaMessage.

fn topic_ptr(&self) -> *mut RDKafkaTopic

Returns a pointer to the message's RDKafkaTopic

fn key_len(&self) -> usize

Returns the length of the key field of the message.

fn payload_len(&self) -> usize

Returns the length of the payload field of the message.

fn detach(&self) -> OwnedMessage

Clones the content of the BorrowedMessage and returns an OwnedMessage, that can outlive the consumer. This operation requires memory allocation and can be expensive.