Trait rdkafka::message::FromBytes

pub trait FromBytes {
    type Error;
    fn from_bytes(_: &[u8]) -> Result<&Self, Self::Error>;
}

Given a reference to a byte array, returns a different view of the same data. No copy of the data should be performed.

Associated Types

type Error

The error type that will be used whenever the conversion fails.

Required Methods

fn from_bytes(_: &[u8]) -> Result<&Self, Self::Error>

Tries to convert the provided byte slice into a different type.