Trait rdkafka::message::Message

pub trait Message {
    fn key(&self) -> Option<&[u8]>;

    fn payload(&self) -> Option<&[u8]>;

    fn topic(&self) -> &str;

    fn partition(&self) -> i32;

    fn offset(&self) -> i64;

    fn timestamp(&self) -> Timestamp;

    fn payload_view<P: ?Sized + FromBytes>(
        &self
    ) -> Option<Result<&P, P::Error>> { ... }

    fn key_view<K: ?Sized + FromBytes>(&self) -> Option<Result<&K, K::Error>> { ... }
}

The Message trait provides access to the fields of a generic Kafka message.

Required Methods

fn key(&self) -> Option<&[u8]>

Returns the key of the message, or None if there is no key.

fn payload(&self) -> Option<&[u8]>

Returns the payload of the message, or None if there is no payload.

fn topic(&self) -> &str

Returns the source topic of the message.

fn partition(&self) -> i32

Returns the partition number where the message is stored.

fn offset(&self) -> i64

Returns the offset of the message.

fn timestamp(&self) -> Timestamp

Returns the message timestamp for a consumed message if available.

Provided Methods

fn payload_view<P: ?Sized + FromBytes>(&self) -> Option<Result<&P, P::Error>>

Converts the raw bytes of the payload to a reference of the specified type, that points to the same data inside the message and without performing any memory allocation

fn key_view<K: ?Sized + FromBytes>(&self) -> Option<Result<&K, K::Error>>

Converts the raw bytes of the key to a reference of the specified type, that points to the same data inside the message and without performing any memory allocation