Trait rdkafka::message::ToBytes

pub trait ToBytes {
    fn to_bytes(&self) -> &[u8];
}

A cheap conversion from typed data to a byte slice.

Given some data, returns the byte representation of that data. No copy of the data should be performed.

See also the FromBytes trait.

Required methods

fn to_bytes(&self) -> &[u8]

Converts the provided data to bytes.