Struct rdkafka::producer::base_producer::BaseProducer

pub struct BaseProducer<C: ProducerContext> { /* fields omitted */ }

Low level Kafka producer.

The BaseProducer needs to be polled at regular intervals in order to serve queued delivery report callbacks (for more information, refer to the module-level documentation. This producer can be cheaply cloned to create a new reference to the same underlying producer.

Methods

impl<C: ProducerContext> BaseProducer<C>

pub fn poll<T: Into<Option<Duration>>>(&self, timeout: T) -> i32

Polls the producer. Regular calls to poll are required to process the events and execute the message delivery callbacks.

pub fn send_copy<P: ?Sized, K: ?Sized>(
&self,
topic_name: &str,
partition: Option<i32>,
payload: Option<&P>,
key: Option<&K>,
delivery_opaque: C::DeliveryOpaque,
timestamp: Option<i64>
) -> KafkaResult<()> where
K: ToBytes,
P: ToBytes,

Sends a copy of the payload and key provided to the specified topic. When no partition is specified the underlying Kafka library picks a partition based on the key. If no key is specified, a random partition will be used. Note that some errors will cause an error to be returned straight-away, such as partition not defined, while others will be returned in the delivery callback. To correctly handle errors, the delivery callback should be implemented.

pub fn flush<T: Into<Option<Duration>>>(&self, timeout: T)

Flushes the producer. Should be called before termination.

pub fn in_flight_count(&self) -> i32

Returns the number of messages waiting to be sent, or send but not acknowledged yet.