Struct rdkafka::producer::base_producer::ThreadedProducer

#[must_use = "The threaded producer will stop immediately if unused"]pub struct ThreadedProducer<C> where
    C: ProducerContext + 'static,  { /* fields omitted */ }

A low-level Kafka producer with a separate thread for event handling.

The ThreadedProducer is a BaseProducer with a separate thread dedicated to calling poll at regular intervals in order to execute any queued events, such as delivery notifications. The thread will be automatically stopped when the producer is dropped.

Implementations

impl<C> ThreadedProducer<C> where
C: ProducerContext + 'static,

pub fn send<'a, K: ?Sized, P: ?Sized>(
&self,
record: BaseRecord<'a, K, P, C::DeliveryOpaque>
) -> Result<(), (KafkaError, BaseRecord<'a, K, P, C::DeliveryOpaque>)> where
K: ToBytes,
P: ToBytes,

Sends a message to Kafka.

See the documentation for BaseProducer::send for details.

pub fn poll<T: Into<Timeout>>(&self, timeout: T)

Polls the internal producer.

This is not normally required since the ThreadedProducer has a thread dedicated to calling poll regularly.