Struct rdkafka::producer::base_producer::ThreadedProducer

#[must_use = "The threaded producer will stop immediately if unused"]pub struct ThreadedProducer<C: ProducerContext + 'static> { /* fields omitted */ }

A low-level Kafka producer with a separate thread for event handling.

The ThreadedProducer is a BaseProducer with a separate thread dedicated to calling poll at regular intervals in order to execute any queued events, such as delivery notifications. The thread will be automatically stopped when the producer is dropped.

Implementations

impl<C: ProducerContext + 'static> ThreadedProducer<C>

pub fn send<'a, K: ?Sized, P: ?Sized>(
&self,
record: BaseRecord<'a, K, P, C::DeliveryOpaque>
) -> Result<(), (KafkaError, BaseRecord<'a, K, P, C::DeliveryOpaque>)> where
K: ToBytes,
P: ToBytes,

Sends a message to Kafka.

See the documentation for BaseProducer::send for details.

pub fn poll<T: Into<Timeout>>(&self, timeout: T)

Polls the internal producer.

This is not normally required since the ThreadedProducer has a thread dedicated to calling poll regularly.

pub fn flush<T: Into<Timeout>>(&self, timeout: T)

Flushes any pending messages.

This method should be called before termination to ensure delivery of all enqueued messages.

pub fn in_flight_count(&self) -> i32

Returns the number of messages that are either waiting to be sent or are sent but are waiting to be acknowledged.

pub fn client(&self) -> &Client<C>

Returns the Client underlying this producer.