Struct rdkafka::producer::future_producer::FutureProducer

#[must_use = "Producer polling thread will stop immediately if unused"]

pub struct FutureProducer<C: ClientContext + 'static = DefaultClientContext> { /* fields omitted */ }

A producer that returns a Future for every message being produced.

Since message production in rdkafka is asynchronous, the called cannot immediately know if the delivery of the message was successful or not. The FutureProducer provides this information in a Future, that will be completed once the information becomes available. This producer has an internal polling thread and as such it doesn't need to be polled. It can be cheaply cloned to get a reference to the same underlying producer. The internal will be terminated once the the FutureProducer goes out of scope.

Methods

impl<C: ClientContext + 'static> FutureProducer<C>

pub fn send_copy<P: ?Sized, K: ?Sized>(
&self,
topic: &str,
partition: Option<i32>,
payload: Option<&P>,
key: Option<&K>,
timestamp: Option<i64>,
block_ms: i64
) -> DeliveryFuture where
K: ToBytes,
P: ToBytes,

Sends a copy of the payload and key provided to the specified topic. When no partition is specified the underlying Kafka library picks a partition based on the key, or a random one if the key is not specified. Returns a DeliveryFuture that will eventually contain the result of the send. The block_ms parameter will control for how long the producer is allowed to block if the queue is full. Set it to -1 to block forever, or 0 to never block. If block_ms is reached and the queue is still full, a RDKafkaError::QueueFull will be reported in the DeliveryFuture.

pub fn send_copy_result<P: ?Sized, K: ?Sized>(
&self,
topic: &str,
partition: Option<i32>,
payload: Option<&P>,
key: Option<&K>,
timestamp: Option<i64>
) -> KafkaResult<DeliveryFuture> where
K: ToBytes,
P: ToBytes,

Sends a copy of the payload and key provided to the specified topic. This works the same way as send_copy, the only difference is that it returns an error if enqueuing fails.

pub fn poll<T: Into<Option<Duration>>>(&self, timeout: T)

Polls the internal producer. This is not normally required since the ThreadedProducer had a thread dedicated to calling poll regularly.

pub fn flush<T: Into<Option<Duration>>>(&self, timeout: T)

Flushes the producer. Should be called before termination.

pub fn in_flight_count(&self) -> i32

Returns the number of messages waiting to be sent, or send but not acknowledged yet.