Struct rdkafka::producer::future_producer::FutureProducerContext

pub struct FutureProducerContext<C: ClientContext + 'static> { /* fields omitted */ }

The ProducerContext used by the FutureProducer.

This context will use a Future as its DeliveryOpaque and will complete the future when the message is delivered (or failed to).