Struct rdkafka::producer::FutureProducer

#[must_use = "Producer polling thread will stop immediately if unused"]

pub struct FutureProducer<C: Context + 'static> { /* fields omitted */ }

A producer with an internal running thread. This producer doesn't need to be polled. It can be cheaply cloned to get a reference to the same underlying producer. The internal thread can be terminated with the stop method or moving the FutureProducer out of scope.

Methods

impl<C: Context + 'static> FutureProducer<C>

fn send_copy<P: ?Sized, K: ?Sized>(
&self,
topic_name: &str,
partition: Option<i32>,
payload: Option<&P>,
key: Option<&K>,
timestamp: Option<i64>
) -> KafkaResult<DeliveryFuture> where
K: ToBytes,
P: ToBytes,

Sends a copy of the payload and key provided to the specified topic. When no partition is specified the underlying Kafka library picks a partition based on the key. Returns a DeliveryFuture or an error.

fn stop(&self)

Stops the internal polling thread. The thread can also be stopped by moving the FutureProducer out of scope.