Trait rdkafka::producer::ProducerContext

pub trait ProducerContext: Context {
    type DeliveryContext: Send + Sync;
    fn delivery(&self, _: DeliveryReport, _: Self::DeliveryContext);
}

A ProducerContext is a Context specific for producers. It can be used to store user-specified callbacks, such as delivery.

Associated Types

type DeliveryContext: Send + Sync

A DeliveryContext is a user-defined structure that will be passed to the producer when producing a message, and returned to the delivery method once the message has been delivered, or failed to.

Required Methods

fn delivery(&self, _: DeliveryReport, _: Self::DeliveryContext)

This method will be called once the message has been delivered (or failed to). The DeliveryContext will be the one provided by the user when calling send.