Struct rdkafka::statistics::Statistics

pub struct Statistics {
    pub name: String,
    pub client_id: String,
    pub client_type: String,
    pub ts: i64,
    pub time: i64,
    pub replyq: i64,
    pub msg_cnt: i64,
    pub msg_size: i64,
    pub msg_max: i64,
    pub msg_size_max: i64,
    pub tx: i64,
    pub tx_bytes: i64,
    pub rx: i64,
    pub rx_bytes: i64,
    pub txmsgs: i64,
    pub txmsg_bytes: i64,
    pub rxmsgs: i64,
    pub rxmsg_bytes: i64,
    pub simple_cnt: i64,
    pub metadata_cache_cnt: i64,
    pub brokers: HashMap<String, Broker>,
    pub topics: HashMap<String, Topic>,
    pub cgrp: Option<ConsumerGroup>,
    pub eos: Option<ExactlyOnceSemantics>,
}

Overall statistics.

Fields
name: String

The name of the librdkafka handle.

client_id: String

The configured client.id.

client_type: String

The instance type (producer or consumer).

ts: i64

The current value of librdkafka's internal monotonic clock, in

time: i64

Wall clock time, in seconds since the Unix epoch.

replyq: i64

The number of operations (callbacks, events, etc.) waiting in queue.

msg_cnt: i64

The current number of messages in producer queues.

msg_size: i64

The current total size of messages in producer queues.

msg_max: i64

The maximum number of messages allowed in the producer queues.

msg_size_max: i64

The maximum total size of messages allowed in the producer queues.

tx: i64

The total number of requests sent to brokers.

tx_bytes: i64

The total number of bytes transmitted to brokers.

rx: i64

The total number of responses received from brokers.

rx_bytes: i64

The total number of bytes received from brokers.

txmsgs: i64

The total number of messages transmitted (produced) to brokers.

txmsg_bytes: i64

The total number of bytes transmitted (produced) to brokers.

rxmsgs: i64

The total number of messages consumed from brokers, not including ignored messages.

rxmsg_bytes: i64

The total number of bytes (including framing) consumed from brokers.

simple_cnt: i64

Internal tracking of legacy vs. new consumer API state.

metadata_cache_cnt: i64

Number of topics in the metadata cache.

brokers: HashMap<String, Broker>

Per-broker statistics.

topics: HashMap<String, Topic>

Per-topic statistics.

cgrp: Option<ConsumerGroup>

Consumer group statistics.

eos: Option<ExactlyOnceSemantics>

Exactly-once semantics and idempotent producer statistics.