Enum rdkafka::topic_partition_list::Offset

pub enum Offset {
    Beginning,
    End,
    Stored,
    Invalid,
    Offset(i64),
}

A librdkafka offset.

Variants
Beginning

Start consuming from the beginning of the partition.

End

Start consuming from the end of the partition.

Stored

Start consuming from the stored offset.

Invalid

Offset not assigned or invalid.

Offset(i64)

A specific offset to consume from.

Methods

impl Offset

pub fn from_raw(raw_offset: i64) -> Offset

Converts the integer representation of an offset use by librdkafka to an Offset.

pub fn to_raw(&self) -> i64

Converts the Offset to the internal integer representation used by librdkafka.