Module rdkafka::util

Utility functions and types.

Structs

NaiveRuntimenaive-runtime

An AsyncRuntime implementation backed by the executor in the futures_executor crate.

TokioRuntimetokio

An AsyncRuntime implementation backed by Tokio.

Enums

Timeout

Specifies a timeout for a Kafka operation.

Traits

AsyncRuntime

An abstraction over asynchronous runtimes.

IntoOpaque

Converts Rust data to and from raw pointers.

Functions

cstr_to_owned⚠

Converts a C string into a String.

current_time_millis

Returns the current time in milliseconds since the Unix epoch.

get_rdkafka_version

Returns a tuple representing the version of librdkafka in hexadecimal and string format.

millis_to_epoch

Converts the given time to the number of milliseconds since the Unix epoch.

Type Definitions

DefaultRuntime

The default AsyncRuntime used when one is not explicitly specified.