Module rdkafka::util

Utility functions

Traits

IntoOpaque
A trait for the conversion of Rust data to raw pointers. This conversion is used to pass opaque objects to the C library and vice versa.

Functions

bytes_cstr_to_owned⚠
Converts a byte array representing a C string into a String.

cstr_to_owned⚠
Converts a C string into a String.

current_time_millis
Returns the current time in millis since unix epoch.

duration_to_millis
Converts a Duration into milliseconds

get_rdkafka_version
Return a tuple representing the version of librdkafka in hexadecimal and string format.

millis_to_epoch
Converts the given time to milliseconds since unix epoch.